requirements.txt
ollama==0.4.4
openai==1.57.2
phidata==2.6.6
pydantic==2.10.3
python-dotenv==1.0.1
yfinance==0.2.51
pydantic_ai==0.0.15
langchain-ollama==0.2.2
langchain-community==0.3.13
pip-system-certs~=4.0
certifi~=2024.12.14
langchain_chroma~=0.1.4
prettyprint~=0.1.5
beautifulsoup4~=4.12.3
